requirements.txt
pydantic_settings
tqdm
loguru
colorama
jedi
python-iso639
python-dotenv
prettytable
tiktoken
openai